Rename opts files and type. (squashed commits)
rename Opts to RuntimeOpts

rename cli-parse to parse-runtime-opts

rename parseCli to parse-compiler-opts

fixes

rename checkOpts to check-compiler-opts

rename checkOpts to check-runtime-opts
