+++ stopify-continuations/src/compiler/check-compiler-opts.ts
