+++ stopify-continuations/src/compiler/tool.ts
@@ -3,7 +3,7 @@
  */
 import * as babel from 'babel-core';
 import * as t from 'babel-types';
-import * as cli from './parseCli';
+import * as cli from './parse-compiler-opts';
 import * as fs from 'fs-extra';
 import { NodePath, Visitor } from 'babel-traverse';
 import callcc from '../callcc/callcc';
