+++ stopify/src/index.ts
@@ -5,7 +5,7 @@ import { generateLineMapping } from './sourceMaps';
 import { RawSourceMap } from 'source-map';
 import { compile } from './compiler/compiler';
 export { compileFunction, compileEval } from './stopify/compileFunction'
-export { CompilerOpts, Opts } from './types';
+export { CompilerOpts, RuntimeOpts } from './types';
 export { stopify as precompiledStopify } from './entrypoints/runtimeOnly';
 export { plugin } from './stopify/stopifyCallCC';
