+++ stopify/src/parse-runtime-opts.ts
@@ -1,6 +1,6 @@
 import * as commander from 'commander';
-import { Opts } from './types';
-import { checkAndFillRuntimeOpts } from './runtime/checkOpts';
+import { RuntimeOpts } from './types';
+import { checkAndFillRuntimeOpts } from './runtime/check-runtime-opts';
 
 commander.option(
   '-y, --yield <interval>',
@@ -46,7 +46,8 @@ commander.option('--local-port <port>',
 
 commander.arguments('<filename>');
 
-export function parseRuntimeOpts(rawArgs: string[]): Opts {
+
+export function parseRuntimeOpts(rawArgs: string[]): RuntimeOpts {
 
   const args = commander.parse(["", "", ...rawArgs]);
